Update directory structure
Updated the directory stucture to be more organized
